move conditions to separate files
