+++ src/utils.ts
@@ -1,4 +1,5 @@
 import { AnyResponse } from '@octokit/rest'
+import { PullRequestInfo } from './models'
 
 export type DeepPartial<T> = { [Key in keyof T]?: DeepPartial<T[Key]>; }
 export type ElementOf<TArray> = TArray extends Array<infer TElement> ? TElement : never
@@ -27,15 +28,15 @@ export function groupByLast<TItem> (
   return groupByLastMap(keyFn, identity, list)
 }
 
-export function groupByLastMap<TItem, TValue> (
-  keyFn: (item: TItem) => string,
+export function groupByLastMap<TItem, TKey extends string, TValue> (
+  keyFn: (item: TItem) => TKey,
   valueFn: (item: TItem) => TValue,
   list: TItem[]
-): { [key: string]: TValue } {
+): { [key in TKey]: TValue } {
   return list.reduce((result, item) => ({
     ...result,
     [keyFn(item)]: valueFn(item)
-  }), {})
+  }), {} as any)
 }
 
 /**
@@ -48,4 +49,60 @@ export function result<TResult = void> (response: AnyResponse): TResult {
     throw new Error(`Response status was ${response.status}`)
   }
   return response.data
+}
+
+export function merge<A, B> (a: A, b: B): A & B {
+  return Object.assign({}, a, b) as any
+}
+
+export function arrayToMap<TKey extends string, TValue, TItem> (
+  arr: Array<TItem>,
+  keyFn: (item: TItem) => TKey,
+  valueFn: (item: TItem) => TValue
+): { [key in TKey]?: TValue } {
+  return arr.reduce((result: { [key in TKey]?: TValue }, item) => merge(result, {
+    [keyFn(item)]: valueFn(item)
+  }), {})
+}
+
+export function or<TValue> (optional: TValue | undefined, defaultValue: TValue): TValue {
+  return optional === undefined
+    ? defaultValue
+    : optional
+}
+
+export function groupByCount<TKey extends string, TItem> (
+  arr: Array<TItem>,
+  keyFn: (item: TItem) => TKey
+): { [key in TKey]?: number } {
+  return arr.reduce((result: { [key in TKey]?: number }, item) => {
+    const key = keyFn(item)
+    const previousValue = result[key]
+    const newValue = or<number>(previousValue, 0) + 1
+    return merge(result, {
+      [key]: newValue + 1
+    })
+  }, {})
+}
+
+export function mapToArray<TKey extends string, TValue> (map: { [key in TKey]?: TValue }) {
+  return Object.entries(map)
+}
+
+export function get<TKey extends string, TValue> (obj: { [key in TKey]: TValue }, key: TKey): TValue {
+  return obj[key]
+}
+
+export function getLatestReviews (pullRequestInfo: PullRequestInfo) {
+  const sortedReviews = pullRequestInfo.reviews.nodes.sort(
+    (a, b) =>
+      new Date(a.submittedAt).getTime() - new Date(b.submittedAt).getTime()
+  )
+  const latestReviewsByUser = groupByLast(
+    review => review.author.login,
+    sortedReviews
+  )
+
+  const latestReviews = Object.values(latestReviewsByUser)
+  return latestReviews
 }
